requirements.txt
fastapi>=0.128.0
uvicorn>=0.40.0
jinja2>=3.1.0
httpx>=0.27.0
python-dotenv>=1.2.0
pydantic>=2.12.0
orjson>=3.10.0
brotli-asgi>=1.4.0
Pillow>=10.0.0
blurhash>=1.1.0
apscheduler>=3.10.0,<4
structlog>=24.0.0
sentry-sdk[fastapi]>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
pytest-httpx>=0.30.0
aiosqlite>=0.20.0
